+++ Src/EGame/Assets/AudioClipAsset.hpp
@@ -9,7 +9,7 @@ namespace eg
 {
 	struct __attribute__ ((__packed__)) AudioClipAssetHeader
 	{
-		uint8_t channelCount;
+		uint32_t channelCount;
 		uint64_t frequency;
 		uint64_t samples;
 	};
